qa-pipeline/requirements.txt
pika
nltk
autocorrect
elasticsearch
numpy
sumy
gensim
bert-extractive-summarizer
redis
